Split UHCI/EHCI port and isoch helpers
