+++ rom/usb/pciusb/ehci/ehciproto.h
@@ -10,6 +10,7 @@ void ehciFreeQHandTDs(struct PCIController *hc, struct EhciQH *eqh);
 void ehciUpdateIntTree(struct PCIController *hc);
 void ehciHandleFinishedTDs(struct PCIController *hc);
 void ehciHandleIsochTDs(struct PCIController *hc);
+void ehciCheckPortStatusChange(struct PCIController *hc);
 void ehciScheduleCtrlTDs(struct PCIController *hc);
 void ehciScheduleIntTDs(struct PCIController *hc);
 void ehciScheduleBulkTDs(struct PCIController *hc);
