+++ rom/usb/pciusb/uhci/uhcichip.c
@@ -86,141 +86,6 @@ void uhciUpdateIntTree(struct PCIController *hc)
     }
 }
 
-static void uhciInsertIsoPTD(struct PCIController *hc, struct PTDNode *ptd, ULONG slot)
-{
-    struct UhciHCPrivate *uhcihcp = (struct UhciHCPrivate *)hc->hc_CPrivate;
-    struct PTDNode *head = uhcihcp->uhc_IsoHead[slot];
-    struct PTDNode *tail = uhcihcp->uhc_IsoTail[slot];
-    struct UhciTD *utd = (struct UhciTD *)ptd->ptd_Descriptor;
-    ULONG prev_entry = READMEM32_LE(&uhcihcp->uhc_UhciFrameList[slot]);
-
-    ptd->ptd_NextPTD = NULL;
-    if(!head) {
-        ptd->ptd_NextPhys = READMEM32_LE(&uhcihcp->uhc_UhciFrameList[slot]);
-        WRITEMEM32_LE(&utd->utd_Link, ptd->ptd_NextPhys);
-        uhcihcp->uhc_IsoHead[slot] = uhcihcp->uhc_IsoTail[slot] = ptd;
-        WRITEMEM32_LE(&uhcihcp->uhc_UhciFrameList[slot], ptd->ptd_Phys);
-        CacheClearE(&uhcihcp->uhc_UhciFrameList[slot], sizeof(ULONG), CACRF_ClearD);
-        pciusbUHCIDebug("UHCI", "ISO insert slot=%ld head prev=%08lx new=%08lx\n",
-                        slot, prev_entry, ptd->ptd_Phys);
-    } else {
-        struct UhciTD *tailtd = (struct UhciTD *)tail->ptd_Descriptor;
-        ULONG nextphys = READMEM32_LE(&tailtd->utd_Link);
-
-        ptd->ptd_NextPhys = head->ptd_NextPhys;
-        WRITEMEM32_LE(&utd->utd_Link, nextphys);
-        WRITEMEM32_LE(&tailtd->utd_Link, ptd->ptd_Phys);
-        CacheClearE(tailtd, sizeof(*tailtd), CACRF_ClearD);
-        tail->ptd_NextPTD = ptd;
-        uhcihcp->uhc_IsoTail[slot] = ptd;
-        pciusbUHCIDebug("UHCI", "ISO insert slot=%ld tail prev=%08lx new=%08lx\n",
-                        slot, nextphys, ptd->ptd_Phys);
-    }
-
-    CacheClearE(utd, sizeof(*utd), CACRF_ClearD);
-}
-
-static void uhciUnlinkIsoPTD(struct PCIController *hc, struct PTDNode *ptd)
-{
-    struct UhciHCPrivate *uhcihcp = (struct UhciHCPrivate *)hc->hc_CPrivate;
-    ULONG slot = ptd->ptd_FrameIdx & (UHCI_FRAMELIST_SIZE - 1);
-    struct PTDNode *prev = NULL;
-    struct PTDNode *head = uhcihcp->uhc_IsoHead[slot];
-
-    while(head && head != ptd) {
-        prev = head;
-        head = head->ptd_NextPTD;
-    }
-
-    if(!head)
-        return;
-
-    struct PTDNode *next = head->ptd_NextPTD;
-
-    if(prev) {
-        struct UhciTD *prevtd = (struct UhciTD *)prev->ptd_Descriptor;
-        WRITEMEM32_LE(&prevtd->utd_Link, next ? next->ptd_Phys : head->ptd_NextPhys);
-        CacheClearE(prevtd, sizeof(*prevtd), CACRF_ClearD);
-        prev->ptd_NextPTD = next;
-    } else {
-        if(next)
-            next->ptd_NextPhys = head->ptd_NextPhys;
-
-        WRITEMEM32_LE(&uhcihcp->uhc_UhciFrameList[slot], next ? next->ptd_Phys : head->ptd_NextPhys);
-        CacheClearE(&uhcihcp->uhc_UhciFrameList[slot], sizeof(ULONG), CACRF_ClearD);
-        uhcihcp->uhc_IsoHead[slot] = next;
-    }
-
-    if(!next)
-        uhcihcp->uhc_IsoTail[slot] = prev;
-
-    head->ptd_NextPTD = NULL;
-}
-
-static struct PTDNode *uhciNextIsoPTD(struct RTIsoNode *rtn)
-{
-    if (!rtn || !rtn->rtn_PTDs || !rtn->rtn_PTDCount)
-        return NULL;
-
-    for (UWORD offset = 0; offset < rtn->rtn_PTDCount; offset++) {
-        UWORD idx = (rtn->rtn_NextPTD + offset) % rtn->rtn_PTDCount;
-        struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-
-        if (!ptd)
-            continue;
-
-        if (!(ptd->ptd_Flags & (PTDF_ACTIVE | PTDF_BUFFER_VALID))) {
-            rtn->rtn_NextPTD = (idx + 1) % rtn->rtn_PTDCount;
-            return ptd;
-        }
-    }
-
-    return NULL;
-}
-
-void uhciCheckPortStatusChange(struct PCIController *hc)
-{
-    struct PCIUnit *unit = hc->hc_Unit;
-    UWORD oldval;
-    UWORD hciport;
-
-    // check for port status change for UHCI and frame rollovers
-
-    for(hciport = 0; hciport < 2; hciport++) {
-        UWORD portreg;
-        UWORD idx = hc->hc_PortNum[hciport];
-        // don't pay attention to UHCI port changes when pwned by EHCI
-        if(unit->hu_PortOwner[idx] == HCITYPE_UHCI) {
-            portreg = hciport ? UHCI_PORT2STSCTRL : UHCI_PORT1STSCTRL;
-            oldval = READIO16_LE(hc->hc_RegBase, portreg);
-            if(oldval & UHPF_ENABLECHANGE) {
-                pciusbUHCIDebug("UHCI", "Port %ld (%ld) Enable changed\n", idx, hciport);
-                hc->hc_PortChangeMap[hciport] |= UPSF_PORT_ENABLE;
-            }
-            if(oldval & UHPF_CONNECTCHANGE) {
-                pciusbUHCIDebug("UHCI", "Port %ld (%ld) Connect changed\n", idx, hciport);
-                hc->hc_PortChangeMap[hciport] |= UPSF_PORT_CONNECTION;
-                if(!(oldval & UHPF_PORTCONNECTED)) {
-                    if(unit->hu_PortMap20[idx]) {
-                        pciusbUHCIDebug("UHCI", "UHCI: Transferring Port %ld back to EHCI\n", idx);
-                        unit->hu_PortOwner[idx] = HCITYPE_EHCI;
-                    }
-                }
-            }
-            if(oldval & UHPF_RESUMEDTX) {
-                pciusbUHCIDebug("UHCI", "Port %ld (%ld) Resume changed\n", idx, hciport);
-                hc->hc_PortChangeMap[hciport] |= UPSF_PORT_SUSPEND|UPSF_PORT_ENABLE;
-                oldval &= ~UHPF_RESUMEDTX;
-            }
-            if(hc->hc_PortChangeMap[hciport]) {
-                unit->hu_RootPortChanges |= 1UL<<(idx+1);
-                /*KPRINTF(10, ("Port %ld (%ld) contributes %04lx to portmap %04lx\n",
-                             idx, hciport, hc->hc_PortChangeMap[hciport], unit->hu_RootPortChanges));*/
-            }
-            WRITEIO16_LE(hc->hc_RegBase, portreg, oldval);
-        }
-    }
-}
 
 void uhciHandleFinishedTDs(struct PCIController *hc)
 {
@@ -815,44 +680,6 @@ void uhciScheduleIntTDs(struct PCIController *hc)
     }
 }
 
-void uhciScheduleIsoTDs(struct PCIController *hc)
-{
-    struct PCIUnit *unit = hc->hc_Unit;
-    struct IOUsbHWReq *ioreq;
-    UWORD devadrep;
-
-    pciusbUHCIDebug("UHCI", "Scheduling new ISO transfers...\n");
-    ioreq = (struct IOUsbHWReq *)hc->hc_IsoXFerQueue.lh_Head;
-    while (((struct Node *)ioreq)->ln_Succ) {
-        devadrep = (ioreq->iouh_DevAddr << 5) + ioreq->iouh_Endpoint +
-                   ((ioreq->iouh_Dir == UHDIR_IN) ? 0x10 : 0);
-        if (unit->hu_DevBusyReq[devadrep]) {
-            ioreq = (struct IOUsbHWReq *)((struct Node *)ioreq)->ln_Succ;
-            continue;
-        }
-
-        struct RTIsoNode *rtn = pciusbAllocStdIsoNode(hc, ioreq);
-        if (!rtn)
-            break;
-
-        Remove(&ioreq->iouh_Req.io_Message.mn_Node);
-
-        if (uhciInitIsochIO(hc, rtn) != RC_OK) {
-            pciusbFreeStdIsoNode(hc, rtn);
-            ioreq->iouh_Req.io_Error = UHIOERR_HOSTERROR;
-            ReplyMsg(&ioreq->iouh_Req.io_Message);
-            ioreq = (struct IOUsbHWReq *)hc->hc_IsoXFerQueue.lh_Head;
-            continue;
-        }
-
-        uhciQueueIsochIO(hc, rtn);
-        AddTail((struct List *)&hc->hc_RTIsoHandlers, (struct Node *)&rtn->rtn_Node);
-        unit->hu_DevBusyReq[devadrep] = ioreq;
-        uhciStartIsochIO(hc, rtn);
-
-        ioreq = (struct IOUsbHWReq *)hc->hc_IsoXFerQueue.lh_Head;
-    }
-}
 
 void uhciScheduleBulkTDs(struct PCIController *hc)
 {
@@ -1417,653 +1244,4 @@ void uhciFree(struct PCIController *hc, struct PCIUnit *hu)
     }
 
     pciusbUHCIDebug("UHCI", "Shutting down UHCI done.\n");
-}
-
-/* ** Root hub support functions ** */
-
-BOOL uhciSetFeature(struct PCIUnit *unit, struct PCIController *hc, UWORD hciport, UWORD idx, UWORD val, WORD *retval)
-{
-    UWORD portreg = hciport ? UHCI_PORT2STSCTRL : UHCI_PORT1STSCTRL;
-    ULONG oldval = READIO16_LE(hc->hc_RegBase, portreg) & ~(UHPF_ENABLECHANGE|UHPF_CONNECTCHANGE); // these are clear-on-write!
-    ULONG newval = oldval;
-    ULONG cnt;
-    BOOL cmdgood = FALSE;
-
-    pciusbUHCIDebug("UHCI", "%s(0x%p, 0x%p, %04x, %04x, %04x, 0x%p)\n", __func__, unit, hc, hciport, idx, val, retval);
-
-    switch(val) {
-    /* case UFS_PORT_CONNECTION: not possible */
-    case UFS_PORT_ENABLE:
-        pciusbUHCIDebug("UHCI", "Enabling Port (%s)\n", newval & UHPF_PORTENABLE ? "already" : "ok");
-        newval |= UHPF_PORTENABLE;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_SUSPEND:
-        newval |= UHPF_PORTSUSPEND;
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_SUSPEND; // manually fake suspend change
-        cmdgood = TRUE;
-        break;
-
-    /* case UFS_PORT_OVER_CURRENT: not possible */
-    case UFS_PORT_RESET:
-        pciusbUHCIDebug("UHCI", "Resetting Port (%s)\n", newval & UHPF_PORTRESET ? "already" : "ok");
-
-        // this is an ugly blocking workaround to the inability of UHCI to clear reset automatically
-        newval &= ~(UHPF_PORTSUSPEND|UHPF_PORTENABLE);
-        newval |= UHPF_PORTRESET;
-        WRITEIO16_LE(hc->hc_RegBase, portreg, newval);
-        uhwDelayMS(25, unit->hu_TimerReq);
-        newval = READIO16_LE(hc->hc_RegBase, portreg) & ~(UHPF_ENABLECHANGE|UHPF_CONNECTCHANGE|UHPF_PORTSUSPEND|UHPF_PORTENABLE);
-        pciusbUHCIDebug("UHCI", "Reset=%s\n", newval & UHPF_PORTRESET ? "GOOD" : "BAD!");
-        // like windows does it
-        newval &= ~UHPF_PORTRESET;
-        WRITEIO16_LE(hc->hc_RegBase, portreg, newval);
-        uhwDelayMicro(50, unit->hu_TimerReq);
-        newval = READIO16_LE(hc->hc_RegBase, portreg) & ~(UHPF_ENABLECHANGE|UHPF_CONNECTCHANGE|UHPF_PORTSUSPEND);
-        pciusbUHCIDebug("UHCI", "Reset=%s\n", newval & UHPF_PORTRESET ? "BAD!" : "GOOD");
-        newval &= ~(UHPF_PORTSUSPEND|UHPF_PORTRESET);
-        newval |= UHPF_PORTENABLE;
-        WRITEIO16_LE(hc->hc_RegBase, portreg, newval);
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_RESET|UPSF_PORT_ENABLE; // manually fake reset change
-
-        cnt = 100;
-        do {
-            uhwDelayMS(1, unit->hu_TimerReq);
-            newval = READIO16_LE(hc->hc_RegBase, portreg);
-        } while(--cnt && (!(newval & UHPF_PORTENABLE)));
-        if(cnt) {
-            pciusbUHCIDebug("UHCI", "Enabled after %ld ticks\n", 100-cnt);
-        } else {
-            pciusbWarn("UHCI", "Port refuses to be enabled!\n");
-            *retval = UHIOERR_HOSTERROR;
-            return TRUE;
-        }
-        // make enumeration possible
-        unit->hu_DevControllers[0] = hc;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_POWER:
-        pciusbUHCIDebug("UHCI", "Powering Port\n");
-        // ignore for UHCI, is always powered
-        cmdgood = TRUE;
-        break;
-
-        /* case UFS_PORT_LOW_SPEED: not possible */
-        /* case UFS_C_PORT_CONNECTION:
-        case UFS_C_PORT_ENABLE:
-        case UFS_C_PORT_SUSPEND:
-        case UFS_C_PORT_OVER_CURRENT:
-        case UFS_C_PORT_RESET: */
-    }
-    if(cmdgood) {
-        pciusbUHCIDebug("UHCI", "Port %ld SET_FEATURE %04lx->%04lx\n", idx, oldval, newval);
-        WRITEIO16_LE(hc->hc_RegBase, portreg, newval);
-    }
-    return cmdgood;
-}
-
-BOOL uhciClearFeature(struct PCIUnit *unit, struct PCIController *hc, UWORD hciport, UWORD idx, UWORD val, WORD *retval)
-{
-    UWORD portreg = hciport ? UHCI_PORT2STSCTRL : UHCI_PORT1STSCTRL;
-    ULONG oldval = READIO16_LE(hc->hc_RegBase, portreg) & ~(UHPF_ENABLECHANGE|UHPF_CONNECTCHANGE); // these are clear-on-write!
-    ULONG newval = oldval;
-    BOOL cmdgood = FALSE;
-
-    pciusbUHCIDebug("UHCI", "%s(0x%p, 0x%p, %04x, %04x, %04x, 0x%p)\n", __func__, unit, hc, hciport, idx, val, retval);
-
-    switch(val) {
-    case UFS_PORT_ENABLE:
-        pciusbUHCIDebug("UHCI", "Disabling Port (%s)\n", newval & UHPF_PORTENABLE ? "ok" : "already");
-        newval &= ~UHPF_PORTENABLE;
-        cmdgood = TRUE;
-        // disable enumeration
-        unit->hu_DevControllers[0] = NULL;
-        break;
-
-    case UFS_PORT_SUSPEND:
-        newval &= ~UHPF_PORTSUSPEND;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_PORT_POWER: // ignore for UHCI, there's no power control here
-        pciusbUHCIDebug("UHCI", "Disabling Power\n");
-        pciusbUHCIDebug("UHCI", "Disabling Port (%s)\n", newval & UHPF_PORTENABLE ? "ok" : "already");
-        newval &= ~UHPF_PORTENABLE;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_CONNECTION:
-        newval |= UHPF_CONNECTCHANGE; // clear-on-write!
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_CONNECTION;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_ENABLE:
-        newval |= UHPF_ENABLECHANGE; // clear-on-write!
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_ENABLE;
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_SUSPEND: // ignore for UHCI, there's no bit indicating this
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_SUSPEND; // manually fake suspend change clearing
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_OVER_CURRENT: // ignore for UHCI, there's no bit indicating this
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_OVER_CURRENT; // manually fake over current clearing
-        cmdgood = TRUE;
-        break;
-
-    case UFS_C_PORT_RESET: // ignore for UHCI, there's no bit indicating this
-        hc->hc_PortChangeMap[hciport] &= ~UPSF_PORT_RESET; // manually fake reset change clearing
-        cmdgood = TRUE;
-        break;
-    }
-    if(cmdgood) {
-        pciusbUHCIDebug("UHCI", "Port %ld CLEAR_FEATURE %04lx->%04lx\n", idx, oldval, newval);
-        WRITEIO16_LE(hc->hc_RegBase, portreg, newval);
-        if(hc->hc_PortChangeMap[hciport]) {
-            unit->hu_RootPortChanges |= 1UL<<(idx + 1);
-        } else {
-            unit->hu_RootPortChanges &= ~(1UL<<(idx + 1));
-        }
-    }
-    return cmdgood;
-}
-
-BOOL uhciGetStatus(struct PCIController *hc, UWORD *mptr, UWORD hciport, UWORD idx, WORD *retval)
-{
-    UWORD portreg = hciport ? UHCI_PORT2STSCTRL : UHCI_PORT1STSCTRL;
-    UWORD oldval = READIO16_LE(hc->hc_RegBase, portreg);
-
-    pciusbUHCIDebug("UHCI", "%s(0x%p, 0x%p, %04x, %04x, 0x%p)\n", __func__, hc, mptr, hciport, idx, retval);
-
-    *mptr = AROS_WORD2LE(UPSF_PORT_POWER);
-
-    if(oldval & UHPF_PORTCONNECTED) *mptr |= AROS_WORD2LE(UPSF_PORT_CONNECTION);
-    if(oldval & UHPF_PORTENABLE) *mptr |= AROS_WORD2LE(UPSF_PORT_ENABLE);
-    if(oldval & UHPF_LOWSPEED) *mptr |= AROS_WORD2LE(UPSF_PORT_LOW_SPEED);
-    if(oldval & UHPF_PORTRESET) *mptr |= AROS_WORD2LE(UPSF_PORT_RESET);
-    if(oldval & UHPF_PORTSUSPEND) *mptr |= AROS_WORD2LE(UPSF_PORT_SUSPEND);
-
-    pciusbUHCIDebug("UHCI", "Port %ld is %s\n", idx, oldval & UHPF_LOWSPEED ? "LOWSPEED" : "FULLSPEED");
-    pciusbUHCIDebug("UHCI", "Port %ld Status %08lx\n", idx, *mptr);
-
-    mptr++;
-    if(oldval & UHPF_ENABLECHANGE) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_ENABLE;
-    }
-    if(oldval & UHPF_CONNECTCHANGE) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_CONNECTION;
-    }
-    if(oldval & UHPF_RESUMEDTX) {
-        hc->hc_PortChangeMap[hciport] |= UPSF_PORT_SUSPEND|UPSF_PORT_ENABLE;
-    }
-
-    *mptr = AROS_WORD2LE(hc->hc_PortChangeMap[hciport]);
-    WRITEIO16_LE(hc->hc_RegBase, portreg, oldval);
-
-    pciusbUHCIDebug("UHCI", "Port %ld Change %08lx\n", idx, *mptr);
-
-    return TRUE;
-}
-
-WORD uhciInitIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    UWORD idx;
-    UWORD ptdcount = rtn->rtn_PTDCount;
-
-    pciusbUHCIDebug("UHCI", "%s()\n", __func__);
-
-    if(!rtn->rtn_PTDs || !ptdcount)
-        return UHIOERR_BADPARAMS;
-
-    for(idx = 0; idx < ptdcount; idx++) {
-        if(rtn->rtn_PTDs[idx]) {
-            if(rtn->rtn_PTDs[idx]->ptd_Descriptor)
-                uhciFreeTD(hc, (struct UhciTD *)rtn->rtn_PTDs[idx]->ptd_Descriptor);
-            FreeMem(rtn->rtn_PTDs[idx], sizeof(struct PTDNode));
-            rtn->rtn_PTDs[idx] = NULL;
-        }
-    }
-
-    for(idx = 0; idx < ptdcount; idx++) {
-        struct PTDNode *ptd = AllocMem(sizeof(*ptd), MEMF_CLEAR);
-        struct UhciTD *utd = ptd ? uhciAllocTD(hc) : NULL;
-
-        if(!ptd || !utd) {
-            if(ptd)
-                FreeMem(ptd, sizeof(*ptd));
-
-            for(ULONG freeidx = 0; freeidx < idx; freeidx++) {
-                struct PTDNode *old = rtn->rtn_PTDs[freeidx];
-                if(old->ptd_Descriptor)
-                    uhciFreeTD(hc, (struct UhciTD *)old->ptd_Descriptor);
-                FreeMem(old, sizeof(*old));
-                rtn->rtn_PTDs[freeidx] = NULL;
-            }
-            return UHIOERR_OUTOFMEMORY;
-        }
-
-        ptd->ptd_Descriptor = utd;
-        ptd->ptd_Phys = READMEM32_LE(&utd->utd_Self);
-        rtn->rtn_PTDs[idx] = ptd;
-    }
-
-    rtn->rtn_NextPTD = 0;
-    rtn->rtn_BounceBuffer = NULL;
-
-    return RC_OK;
-}
-
-WORD uhciQueueIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    struct IOUsbHWReq *ioreq = pciusbIsoGetIOReq(rtn);
-    struct PTDNode *ptd = uhciNextIsoPTD(rtn);
-    struct IOUsbHWRTIso *urti = rtn->rtn_RTIso;
-    ULONG interval;
-    BOOL explicit_frame = FALSE;
-
-    pciusbUHCIDebug("UHCI", "%s(0x%p, 0x%p)\n", __func__, hc, rtn);
-    pciusbUHCIDebug("UHCI", "%s: IOReq 0x%p\n", __func__, ioreq);
-
-    if (!ptd)
-        return UHIOERR_NAKTIMEOUT;
-
-    struct IOUsbHWBufferReq *bufreq = &ptd->ptd_BufferReq;
-    *bufreq = rtn->rtn_BufferReq;
-    explicit_frame = (rtn->rtn_Flags & RTISO_FLAG_EXPLICIT_FRAME) != 0;
-
-    if (urti && !explicit_frame)
-        bufreq->ubr_Frame = 0;
-
-    if (urti) {
-        if (ioreq->iouh_Dir == UHDIR_IN) {
-            if (urti->urti_InReqHook)
-                CallHookPkt(urti->urti_InReqHook, rtn, bufreq);
-        } else {
-            if (urti->urti_OutReqHook)
-                CallHookPkt(urti->urti_OutReqHook, rtn, bufreq);
-        }
-        if (bufreq->ubr_Frame)
-            explicit_frame = TRUE;
-    } else {
-        bufreq->ubr_Buffer = ioreq->iouh_Data;
-        bufreq->ubr_Length = ioreq->iouh_Length;
-        bufreq->ubr_Frame = ioreq->iouh_Frame;
-        bufreq->ubr_Flags = 0;
-    }
-
-    if (!bufreq->ubr_Length)
-        bufreq->ubr_Length = ioreq->iouh_Length;
-
-    if (urti) {
-        if (explicit_frame)
-            rtn->rtn_Flags |= RTISO_FLAG_EXPLICIT_FRAME;
-        else
-            rtn->rtn_Flags &= ~RTISO_FLAG_EXPLICIT_FRAME;
-    }
-
-    if (urti && !(rtn->rtn_Flags & RTISO_FLAG_EXPLICIT_FRAME))
-        bufreq->ubr_Frame = 0;
-
-    interval = ioreq->iouh_Interval ? ioreq->iouh_Interval : 1;
-    if (!bufreq->ubr_Frame) {
-        ULONG current_frame;
-        ULONG lead = interval * 8;
-        ULONG next = 0;
-
-        if (lead < 8)
-            lead = 8;
-
-        uhciUpdateFrameCounter(hc);
-        current_frame = hc->hc_FrameCounter;
-
-        if (rtn->rtn_NextFrame > current_frame)
-            next = rtn->rtn_NextFrame;
-        else
-            next = current_frame + lead;
-
-        bufreq->ubr_Frame = next;
-        pciusbUHCIDebug("UHCI", "ISO schedule current=%ld next=%ld lead=%ld interval=%ld\n",
-                        current_frame, bufreq->ubr_Frame, lead, interval);
-    } else {
-        ULONG current_frame;
-        ULONG lead = interval * 8;
-        LONG delta;
-
-        if (lead < 8)
-            lead = 8;
-
-        uhciUpdateFrameCounter(hc);
-        current_frame = hc->hc_FrameCounter;
-        delta = (LONG)bufreq->ubr_Frame - (LONG)current_frame;
-        if (delta <= 0 || delta >= (LONG)(UHCI_FRAMELIST_SIZE - lead)) {
-            bufreq->ubr_Frame = current_frame + lead;
-            pciusbUHCIDebug("UHCI", "ISO schedule adjusted current=%ld next=%ld lead=%ld interval=%ld\n",
-                            current_frame, bufreq->ubr_Frame, lead, interval);
-        } else {
-            pciusbUHCIDebug("UHCI", "ISO schedule current=%ld interval=%ld\n",
-                            bufreq->ubr_Frame, interval);
-        }
-    }
-    if (bufreq->ubr_Frame < hc->hc_FrameCounter) {
-        ULONG lead = interval * 8;
-
-        if (lead < 8)
-            lead = 8;
-
-        bufreq->ubr_Frame = hc->hc_FrameCounter + lead;
-        pciusbUHCIDebug("UHCI", "ISO schedule resync current=%ld next=%ld lead=%ld interval=%ld\n",
-                        hc->hc_FrameCounter, bufreq->ubr_Frame, lead, interval);
-    }
-    rtn->rtn_NextFrame = bufreq->ubr_Frame + interval;
-
-    ptd->ptd_FrameIdx = bufreq->ubr_Frame;
-    ptd->ptd_Flags |= PTDF_BUFFER_VALID;
-    ptd->ptd_BounceBuffer = NULL;
-    rtn->rtn_BufferReq = *bufreq;
-
-    return RC_OK;
-}
-
-void uhciHandleIsochTDs(struct PCIController *hc)
-{
-    struct PCIUnit *unit = hc->hc_Unit;
-    struct RTIsoNode *rtn = (struct RTIsoNode *)hc->hc_RTIsoHandlers.mlh_Head;
-    ULONG current_frame;
-
-    uhciUpdateFrameCounter(hc);
-    current_frame = hc->hc_FrameCounter;
-
-    while(rtn->rtn_Node.mln_Succ) {
-        struct RTIsoNode *next = (struct RTIsoNode *)rtn->rtn_Node.mln_Succ;
-        struct IOUsbHWRTIso *urti = rtn->rtn_RTIso;
-        UWORD idx;
-        UWORD limit = rtn->rtn_PTDCount;
-
-        if(!rtn->rtn_PTDs || !limit) {
-            rtn = (struct RTIsoNode *) rtn->rtn_Node.mln_Succ;
-            continue;
-        }
-
-        for(idx = 0; idx < limit; idx++) {
-            struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-            struct UhciTD *utd;
-            ULONG ctrl;
-            ULONG actual;
-            ULONG interval;
-            BOOL error = FALSE;
-            struct IOUsbHWReq *ioreq = pciusbIsoGetIOReq(rtn);
-
-            if(!ptd || !(ptd->ptd_Flags & PTDF_ACTIVE))
-                continue;
-
-            utd = (struct UhciTD *)ptd->ptd_Descriptor;
-            CacheClearE(utd, sizeof(*utd), CACRF_InvalidateD);
-            ctrl = READMEM32_LE(&utd->utd_CtrlStatus);
-            if(ctrl & UTCF_ACTIVE) {
-                interval = rtn->rtn_IOReq.iouh_Interval ? rtn->rtn_IOReq.iouh_Interval : 1;
-                ULONG timeout_window = interval * 8;
-
-                if (timeout_window < 16)
-                    timeout_window = 16;
-
-                if (current_frame > (ptd->ptd_FrameIdx + timeout_window)) {
-                    error = TRUE;
-                    ioreq->iouh_Req.io_Error = UHIOERR_TIMEOUT;
-                    ioreq->iouh_Actual = 0;
-                    ptd->ptd_BufferReq.ubr_Length = 0;
-                    ptd->ptd_BufferReq.ubr_Frame = ptd->ptd_FrameIdx;
-                    pciusbUHCIDebug("UHCI", "ISO TD timeout ptd=%p frame=%ld current=%ld window=%ld\n",
-                                    ptd, ptd->ptd_FrameIdx, current_frame, timeout_window);
-                    uhciUnlinkIsoPTD(hc, ptd);
-                    if (urti) {
-                        if (ioreq->iouh_Dir == UHDIR_IN) {
-                            if (urti->urti_InDoneHook)
-                                CallHookPkt(urti->urti_InDoneHook, rtn, &ptd->ptd_BufferReq);
-                        } else {
-                            if (urti->urti_OutDoneHook)
-                                CallHookPkt(urti->urti_OutDoneHook, rtn, &ptd->ptd_BufferReq);
-                        }
-                    }
-                    ptd->ptd_Flags &= ~(PTDF_ACTIVE|PTDF_BUFFER_VALID);
-                    if (utd) {
-                        uhciFreeTD(hc, utd);
-                        ptd->ptd_Descriptor = NULL;
-                    }
-                    if (rtn->rtn_StdReq) {
-                        UWORD devadrep = (ioreq->iouh_DevAddr << 5) + ioreq->iouh_Endpoint +
-                                         ((ioreq->iouh_Dir == UHDIR_IN) ? 0x10 : 0);
-                        Remove((struct Node *)&rtn->rtn_Node);
-                        unit->hu_DevBusyReq[devadrep] = NULL;
-                        ReplyMsg(&ioreq->iouh_Req.io_Message);
-                        uhciFreeIsochIO(hc, rtn);
-                        pciusbFreeStdIsoNode(hc, rtn);
-                    } else if (urti) {
-                        UWORD pending = 0;
-                        UWORD target = (rtn->rtn_PTDCount > 1) ? 2 : 1;
-
-                        for (UWORD scan = 0; scan < rtn->rtn_PTDCount; scan++) {
-                            struct PTDNode *ptdscan = rtn->rtn_PTDs[scan];
-                            if (ptdscan && (ptdscan->ptd_Flags & (PTDF_ACTIVE | PTDF_BUFFER_VALID)))
-                                pending++;
-                        }
-
-                        while (pending < target) {
-                            if (uhciQueueIsochIO(hc, rtn) != RC_OK)
-                                break;
-                            uhciStartIsochIO(hc, rtn);
-                            pending++;
-                        }
-                    }
-                    continue;
-                }
-                if (current_frame > ptd->ptd_FrameIdx) {
-                    pciusbUHCIDebug("UHCI", "ISO TD still active ptd=%p frame=%ld current=%ld ctrl=%08lx\n",
-                                    ptd, ptd->ptd_FrameIdx, current_frame, ctrl);
-                }
-                continue;
-            }
-
-            if(ctrl & (UTSF_BITSTUFFERR|UTSF_CRCTIMEOUT|UTSF_BABBLE|UTSF_DATABUFFERERR))
-                error = TRUE;
-
-            actual = (ctrl & UTSM_ACTUALLENGTH) >> UTSS_ACTUALLENGTH;
-            actual = (actual == 0x7ff) ? 0 : actual + 1;
-            ioreq->iouh_Actual = actual;
-            ioreq->iouh_Req.io_Error = error ? UHIOERR_HOSTERROR : UHIOERR_NO_ERROR;
-            ptd->ptd_BufferReq.ubr_Length = actual;
-            ptd->ptd_BufferReq.ubr_Frame = ptd->ptd_FrameIdx;
-
-            if(ptd->ptd_BounceBuffer &&
-                    ptd->ptd_BounceBuffer != ptd->ptd_BufferReq.ubr_Buffer) {
-                usbReleaseBuffer(ptd->ptd_BounceBuffer, ptd->ptd_BufferReq.ubr_Buffer,
-                                 ptd->ptd_BufferReq.ubr_Length, ioreq->iouh_Dir);
-                ptd->ptd_BounceBuffer = NULL;
-            }
-
-            uhciUnlinkIsoPTD(hc, ptd);
-
-            if(urti) {
-                if(ioreq->iouh_Dir == UHDIR_IN) {
-                    if(urti->urti_InDoneHook)
-                        CallHookPkt(urti->urti_InDoneHook, rtn, &ptd->ptd_BufferReq);
-                } else {
-                    if(urti->urti_OutDoneHook)
-                        CallHookPkt(urti->urti_OutDoneHook, rtn, &ptd->ptd_BufferReq);
-                }
-            }
-
-            ptd->ptd_Flags &= ~(PTDF_ACTIVE|PTDF_BUFFER_VALID);
-            uhciFreeTD(hc, utd);
-            ptd->ptd_Descriptor = NULL;
-
-            if (rtn->rtn_StdReq) {
-                UWORD devadrep = (ioreq->iouh_DevAddr << 5) + ioreq->iouh_Endpoint +
-                                 ((ioreq->iouh_Dir == UHDIR_IN) ? 0x10 : 0);
-                Remove((struct Node *)&rtn->rtn_Node);
-                unit->hu_DevBusyReq[devadrep] = NULL;
-                ioreq->iouh_Actual = actual;
-                ReplyMsg(&ioreq->iouh_Req.io_Message);
-                uhciFreeIsochIO(hc, rtn);
-                pciusbFreeStdIsoNode(hc, rtn);
-            } else if (urti) {
-                UWORD pending = 0;
-                UWORD target = (rtn->rtn_PTDCount > 1) ? 2 : 1;
-
-                for (UWORD scan = 0; scan < rtn->rtn_PTDCount; scan++) {
-                    struct PTDNode *ptdscan = rtn->rtn_PTDs[scan];
-                    if (ptdscan && (ptdscan->ptd_Flags & (PTDF_ACTIVE | PTDF_BUFFER_VALID)))
-                        pending++;
-                }
-
-                while (pending < target) {
-                    if (uhciQueueIsochIO(hc, rtn) != RC_OK)
-                        break;
-                    uhciStartIsochIO(hc, rtn);
-                    pending++;
-                }
-            }
-        }
-
-        rtn = next;
-    }
-}
-
-void uhciStartIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    UWORD idx;
-    UWORD limit = rtn->rtn_PTDCount;
-
-    pciusbUHCIDebug("UHCI", "%s()\n", __func__);
-
-    if(!rtn->rtn_PTDs || !limit)
-        return;
-
-    for(idx = 0; idx < limit; idx++) {
-        struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-        struct UhciTD *utd;
-        ULONG ctrlstatus;
-        ULONG token;
-        ULONG slot;
-        ULONG phys;
-        ULONG len;
-        APTR dmabuffer;
-
-        if(!ptd)
-            continue;
-
-        if(!(ptd->ptd_Flags & PTDF_BUFFER_VALID) || (ptd->ptd_Flags & PTDF_ACTIVE))
-            continue;
-
-        if(!ptd->ptd_Descriptor) {
-            ptd->ptd_Descriptor = uhciAllocTD(hc);
-            if(!ptd->ptd_Descriptor)
-                continue;
-            ptd->ptd_Phys = READMEM32_LE(&((struct UhciTD *)ptd->ptd_Descriptor)->utd_Self);
-        }
-
-        dmabuffer = ptd->ptd_BufferReq.ubr_Buffer;
-#if __WORDSIZE == 64
-        if(dmabuffer) {
-            dmabuffer = usbGetBuffer(dmabuffer, ptd->ptd_BufferReq.ubr_Length,
-                                     rtn->rtn_IOReq.iouh_Dir);
-            if(!dmabuffer)
-                continue;
-
-            if(dmabuffer != ptd->ptd_BufferReq.ubr_Buffer &&
-                    rtn->rtn_IOReq.iouh_Dir == UHDIR_OUT) {
-                CopyMem(ptd->ptd_BufferReq.ubr_Buffer, dmabuffer, ptd->ptd_BufferReq.ubr_Length);
-            }
-        }
-#endif
-        ptd->ptd_BounceBuffer = (dmabuffer != ptd->ptd_BufferReq.ubr_Buffer) ? dmabuffer : NULL;
-
-        len = ptd->ptd_BufferReq.ubr_Length;
-        if(len > UHCI_ISO_MAXPKTSIZE)
-            len = UHCI_ISO_MAXPKTSIZE;
-        if(!len)
-            phys = 0;
-        if(len && !dmabuffer) {
-            ptd->ptd_Flags &= ~PTDF_BUFFER_VALID;
-            continue;
-        }
-
-        utd = (struct UhciTD *)ptd->ptd_Descriptor;
-        slot = ptd->ptd_BufferReq.ubr_Frame & (UHCI_FRAMELIST_SIZE - 1);
-        if(len)
-            phys = (ULONG)(IPTR)pciGetPhysical(hc, dmabuffer);
-
-        ctrlstatus = UTCF_ACTIVE | UTCF_ISOCHRONOUS | UTCF_READYINTEN;
-        token = (rtn->rtn_IOReq.iouh_DevAddr << UTTS_DEVADDR) |
-                (rtn->rtn_IOReq.iouh_Endpoint << UTTS_ENDPOINT) |
-                (((len ? (len - 1) : 0x7ff) & 0x7ff) << UTTS_TRANSLENGTH) |
-                ((rtn->rtn_IOReq.iouh_Dir == UHDIR_IN) ? (PID_IN << UTTS_PID) : (PID_OUT << UTTS_PID));
-
-        WRITEMEM32_LE(&utd->utd_CtrlStatus, ctrlstatus);
-        WRITEMEM32_LE(&utd->utd_Token, token);
-        WRITEMEM32_LE(&utd->utd_BufferPtr, phys);
-
-        ptd->ptd_FrameIdx = ptd->ptd_BufferReq.ubr_Frame;
-        ptd->ptd_Length = len;
-        ptd->ptd_Flags |= PTDF_ACTIVE;
-        ptd->ptd_NextPTD = NULL;
-
-        pciusbUHCIDebug("UHCI", "ISO TD ptd=%p utd=%p frame=%ld slot=%ld len=%ld\n",
-                        ptd, utd, ptd->ptd_FrameIdx, slot, len);
-        uhciInsertIsoPTD(hc, ptd, slot);
-    }
-}
-
-void uhciStopIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    UWORD idx;
-    UWORD limit = rtn->rtn_PTDCount;
-
-    pciusbUHCIDebug("UHCI", "%s()\n", __func__);
-
-    if(!rtn->rtn_PTDs || !limit)
-        return;
-
-    for(idx = 0; idx < limit; idx++) {
-        struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-        if(ptd && (ptd->ptd_Flags & PTDF_ACTIVE)) {
-            uhciUnlinkIsoPTD(hc, ptd);
-            ptd->ptd_Flags &= ~(PTDF_ACTIVE|PTDF_BUFFER_VALID);
-        }
-        if(ptd && ptd->ptd_BounceBuffer &&
-                ptd->ptd_BounceBuffer != ptd->ptd_BufferReq.ubr_Buffer) {
-            usbReleaseBuffer(ptd->ptd_BounceBuffer, ptd->ptd_BufferReq.ubr_Buffer,
-                             ptd->ptd_BufferReq.ubr_Length, rtn->rtn_IOReq.iouh_Dir);
-            ptd->ptd_BounceBuffer = NULL;
-        }
-    }
-}
-
-void uhciFreeIsochIO(struct PCIController *hc, struct RTIsoNode *rtn)
-{
-    UWORD idx;
-    UWORD limit = rtn->rtn_PTDCount;
-
-    pciusbUHCIDebug("UHCI", "%s()\n", __func__);
-
-    uhciStopIsochIO(hc, rtn);
-
-    if(!rtn->rtn_PTDs)
-        return;
-
-    for(idx = 0; idx < limit; idx++) {
-        struct PTDNode *ptd = rtn->rtn_PTDs[idx];
-        if(ptd) {
-            if(ptd->ptd_Descriptor)
-                uhciFreeTD(hc, (struct UhciTD *)ptd->ptd_Descriptor);
-            FreeMem(ptd, sizeof(*ptd));
-            rtn->rtn_PTDs[idx] = NULL;
-        }
-    }
-
-    rtn->rtn_BounceBuffer = NULL;
 }
